Move test
- Test in wrong file, Markdown rather than HTML.
